+++ nix/packages/freecad.nix
@@ -17,6 +17,7 @@ let
     "${perSystem.self.freecad-mcp}/share/freecad-mcp/FreeCADMCP"
     inputs.gridfinity
     inputs.curves
+    perSystem.self.bambucad
   ];
 
   darkPack = "${freecad}/share/Gui/PreferencePacks/FreeCAD Dark/FreeCAD Dark.cfg";